Create CVE-2023-23752
This is a conversion of Ruby language to Python of the vulnerability Joomla! v4.2.8 - Unauthenticated information disclosure. This code is in Exploit-DB written in Ruby by ALEXANDRE ZANNI. 

I had some issues with the libraries of Ruby and I decided to convert the code to Python and worked well. Works the same as the Ruby version.

Original version is here: https://www.exploit-db.com/exploits/51334